Thread create updated on windows, and fixed a bunch of warnings.
FossilOrigin-Name: 0a1683426f5a68dc50a50469bc07f44bfbd676a08c99fab608b2e5757e526450

+++ include/zodiac/common.h
@@ -9,9 +9,9 @@
 #include <string.h>
 
 // Common macros
-#define KB(n) ((n) * 1024)
-#define MB(n) (KB((n)) * 1024)
-#define GB(n) (MB((n)) * 1024)
+#define KB(n) ((n) * (uint64_t)1024)
+#define MB(n) (KB((n)) * (uint64_t)1024)
+#define GB(n) (MB((n)) * (uint64_t)1024)
 
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
